userd: Split core into own submodule
