+++ surface-dtx-userd/src/logic/mod.rs
@@ -1,23 +1,20 @@
+mod core;
+use self::core::Core;
+
 mod types;
-use types::Event;
+use self::types::Event;
 
 
-use crate::utils::notify::{Notification, NotificationHandle, Timeout};
 use crate::utils::task::JoinHandleExt;
-
-use std::str::FromStr;
-use std::sync::Arc;
 
 use anyhow::{Context, Result};
 
-use dbus::Message;
 use dbus::message::MatchRule;
-use dbus::nonblock::SyncConnection;
 use dbus_tokio::connection;
 
 use futures::prelude::*;
 
-use tracing::{debug, trace};
+use tracing::trace;
 
 
 pub async fn run() -> Result<()> {
@@ -44,8 +41,16 @@ pub async fn run() -> Result<()> {
             .context("Failed to set up D-Bus connection")?
             .msg_stream();
 
-        while let Some(m) = stream.next().await {
-            core.handle(m).await?;
+        while let Some(mut msg) = stream.next().await {
+            trace!(target: "sdtxu::core", message = ?msg, "message received");
+
+            let msg = msg.as_result().context("D-Bus remote error")?;
+            let evt = Event::try_from_message(msg)?;
+            drop(msg);
+
+            if let Some(evt) = evt {
+                core.handle(evt).await?;
+            }
         }
 
         Ok(())
@@ -68,121 +73,5 @@ pub async fn run() -> Result<()> {
         Ok(result) => result,
         Err(e) if e.is_panic() => std::panic::resume_unwind(e.into_panic()),
         Err(_) => unreachable!("Subtask canceled before completing parent task"),
-    }
-}
-
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum Status {
-    DetachReady,
-    DetachCompleted,
-    DetachAborted,
-    AttachCompleted,
-}
-
-impl FromStr for Status {
-    type Err = anyhow::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "detach-ready"     => Ok(Self::DetachReady),
-            "detach-completed" => Ok(Self::DetachCompleted),
-            "detach-aborted"   => Ok(Self::DetachAborted),
-            "attach-completed" => Ok(Self::AttachCompleted),
-            _ => anyhow::bail!("Invalid detachment state: '{}'", s),
-        }
-    }
-}
-
-impl std::fmt::Display for Status {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Status::DetachReady     => write!(f, "detach-ready"),
-            Status::DetachCompleted => write!(f, "detach-completed"),
-            Status::DetachAborted   => write!(f, "detach-aborted"),
-            Status::AttachCompleted => write!(f, "attach-completed"),
-        }
-    }
-}
-
-
-#[derive(Clone)]
-struct Core {
-    session:      Arc<SyncConnection>,
-    detach_notif: Option<NotificationHandle>,
-}
-
-impl Core {
-    fn new(session: Arc<SyncConnection>) -> Self {
-        Core {
-            session,
-            detach_notif: None,
-        }
-    }
-
-    async fn handle(&mut self, mut message: Message) -> Result<()> {
-        trace!(target: "sdtxu::core", msg = ?message, "message received");
-
-        let m = message.as_result().context("D-Bus remote error")?;
-
-        if let Some(event) = Event::try_from_message(m)? {
-            debug!(target: "sdtxu::core", ?event, "event received");
-
-            // TODO
-        }
-
-        Ok(())
-    }
-
-    #[allow(unused)]
-    async fn notify_detach_ready(&mut self) -> Result<()> {
-        let handle = Notification::create("Surface DTX")
-            .summary("Surface DTX")
-            .body("Clipboard can be detached.")
-            .hint_s("image-path", "input-tablet")
-            .hint_s("category", "device")
-            .hint("urgency", 2)
-            .hint("resident", true)
-            .expires(Timeout::Never)
-            .build()
-            .show(&self.session).await
-            .context("Failed to display notification")?;
-
-        trace!(target: "sdtxu::notify", id = handle.id, ty = "detach-ready",
-               "displaying notification");
-
-        self.detach_notif = Some(handle);
-        Ok(())
-    }
-
-    #[allow(unused)]
-    async fn notify_detach_completed(&mut self) -> Result<()> {
-        if let Some(handle) = self.detach_notif {
-            trace!(target: "sdtxu::notify", id = handle.id, ty = "detach-ready",
-                   "closing notification");
-
-            handle.close(&self.session).await
-                .context("Failed to close notification")?;
-        }
-
-        Ok(())
-    }
-
-    #[allow(unused)]
-    async fn notify_attach_completed(&self) -> Result<()> {
-        let handle = Notification::create("Surface DTX")
-            .summary("Surface DTX")
-            .body("Clipboard attached.")
-            .hint_s("image-path", "input-tablet")
-            .hint_s("category", "device")
-            .hint("transient", true)
-            .build()
-            .show(&self.session).await
-            .context("Failed to display notification")?;
-
-        trace!(target: "sdtxu::notify", id = handle.id, ty = "attach-complete",
-               "displaying notification");
-
-        Ok(())
     }
 }
